refactor: extract openwhyd api v2 to dedicated file
